load translations before the application starts
